# docs/api-gateway-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "api-gateway-service",
    "description": "HuleEdu API Gateway - Secure client-facing API for React frontend integration",
    "contact": {
      "name": "HuleEdu Development Team",
      "url": "https://github.com/huledu/huledu-reboot"
    },
    "license": {
      "name": "MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/batches/{batch_id}/status": {
      "get": {
        "tags": [
          "Status"
        ],
        "summary": "Get Batch Status",
        "description": "Get batch status with strict ownership enforcement.\n\nActs as a pure proxy to the Result Aggregator Service which handles all\nstatus aggregation and any necessary fallback logic internally.",
        "operationId": "get_batch_status_v1_batches__batch_id__status_get",
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Batch Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BatchClientStatus": {
        "type": "string",
        "enum": [
          "available",
          "processing"
        ],
        "title": "BatchClientStatus",
        "description": "Defines the status of a batch as reported to the client."
      },
...
      "BatchStatusResponse": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/BatchClientStatus",
            "description": "The current status of the batch."
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details",
            "description": "The detailed status information."
          }
        },
        "type": "object",
        "required": [
          "status",
          "details"
        ],
        "title": "BatchStatusResponse",
        "description": "Response model for batch status requests."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}